+++ test-parts/09-api_frontend/parts/02-add_all_devices.js
@@ -154,7 +154,7 @@ function getDeviceCreateUrl()
 	{
 		res = apiRequestScript.writeUrl(apiPaths.devicesApi, apiPaths.rioApiSub);
 	}
-	catch(e)
+	catch(createErr)
 	{
 		res = null;
 	}