+++ test-parts/09-api_frontend/parts/04-check_device_list.js
@@ -96,7 +96,7 @@ function retrieveListFromCache()
 	{
 		res = testCacheFile.retrieveList();
 	}
-	catch(catchErr)
+	catch(retErr)
 	{
 		res = null;
 	}
@@ -113,7 +113,7 @@ function retrieveCountFromCache()
 	{
 		res = testCacheFile.countTestDevices();
 	}
-	catch(catchErr)
+	catch(retErr)
 	{
 		res = null;
 	}