+++ test-parts/09-api_frontend/parts/06-check_node_array.js
@@ -116,7 +116,7 @@ function callNodeCache()
 	{
 		res = testCacheFile.retrieveNodeArrayStructure();
 	}
-	catch(catchErr)
+	catch(callErr)
 	{
 		res = null;
 	}