+++ test-parts/09-api_frontend/parts/07-get_device_properties.js
@@ -145,7 +145,7 @@ function callNodeManufacturerPropertyList()
 	{
 		listRes = testCacheFile.readNodePropertyList();
 	}
-	catch(catchErr)
+	catch(callErr)
 	{
 		listRes = null;
 	}