+++ test-parts/09-api_frontend/parts/08-delete_added_devices.js
@@ -72,7 +72,7 @@ function getAddedDeviceCountFromCache()
 	{
 		addRes = testCacheFile.countTestDevices();
 	}
-	catch(catchErr)
+	catch(getCountErr)
 	{
 		addRes = null;
 	}